// src/app/models/Animal.ts
export class Animal {
  id_animal: number | undefined;
  nombre_animal: string | undefined;
  e_conservacion: string | undefined;
  dieta: string | undefined;
  reproduccion: string | undefined;
  adaptacion: string | undefined;
  amenazas: string | undefined;
  id_especie: number | undefined;
  nombre_especie: string | undefined;
  id_zona: number | undefined;
  nombre_zona: string | undefined;
  id_ecosistema: number | undefined;
  nombre_ecosistema: string | undefined;
}